+++ src/ComputerTab.php
@@ -482,4 +482,8 @@ class ComputerTab extends DeviceTab implements Ticketable {
         return  $item->fields['name'] . "/" . $item->fields['p_name'];
     }
 
+    static function getDeviceForeignKeyField(): string
+    {
+        return Computer::getForeignKeyField();
+    }
 }